+++ src/_components/ModelNumber/ModelNumberContainer.jsx
@@ -3,11 +3,11 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import _ from 'lodash';
 
-import AssetMakeComponent from '../../components/AssetMake/AssetMakeComponent'
+import ModelNumberComponent from '../../components/ModelNumber/ModelNumberComponent'
 
 import { loadAssetMakes } from '../../_actions/assetMakes.actions';
 
-class AssetMakeContainer extends React.Component {
+class ModelNumberContainer extends React.Component {
   constructor(props) {
     super(props);
   }
@@ -19,11 +19,11 @@ class AssetMakeContainer extends React.Component {
   }
 
   render(){
-    return <AssetMakeComponent {...this.props}/>
+    return <ModelNumberComponent {...this.props}/>
   }
 
 }
 const mapStateToProps = ({ assetMakesList }) => ({
   assetMakes: assetMakesList
 });
-export default connect(mapStateToProps, {loadAssetMakes})(AssetMakeContainer);
+export default connect(mapStateToProps, {loadAssetMakes})(ModelNumberContainer);
